更新了实体类

+++ creditcard-interface/src/main/java/com/credit/pojo/TbAppCardStatus.java
@@ -10,37 +10,37 @@ import java.util.Date;
 public class TbAppCardStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;//自增id
+    private Long id;//自增id
 
-    private Integer uid;//用户uid
+    private Long uid;//用户uid
 
     private String cardType;//片卡类型
 
     private String cardName;//卡片名称
 
-    private Integer appStatus;//申请中的状态码（1为1级员工查看，2为2级员工查看，3为3级员工查看）
+    private Long appStatus;//申请中的状态码（1为1级员工查看，2为2级员工查看，3为3级员工查看）
 
-    private Integer cardStatus;//卡片状态（1为已申请，2为未通过，3为已通过，4制卡中，5为寄送中，6已签收）
+    private Long cardStatus;//卡片状态（1为已申请，2为未通过，3为已通过，4制卡中，5为寄送中，6已签收）
 
     private String logisticeInfo;//物流信息
 
-    private Integer cradId;//卡号（通过申请后才会有）
+    private Long cradId;//卡号（通过申请后才会有）
 
     private Date appTime;//申请时间
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
 
@@ -60,19 +60,19 @@ public class TbAppCardStatus {
         this.cardName = cardName == null ? null : cardName.trim();
     }
 
-    public Integer getAppStatus() {
+    public Long getAppStatus() {
         return appStatus;
     }
 
-    public void setAppStatus(Integer appStatus) {
+    public void setAppStatus(Long appStatus) {
         this.appStatus = appStatus;
     }
 
-    public Integer getCardStatus() {
+    public Long getCardStatus() {
         return cardStatus;
     }
 
-    public void setCardStatus(Integer cardStatus) {
+    public void setCardStatus(Long cardStatus) {
         this.cardStatus = cardStatus;
     }
 
@@ -84,11 +84,11 @@ public class TbAppCardStatus {
         this.logisticeInfo = logisticeInfo == null ? null : logisticeInfo.trim();
     }
 
-    public Integer getCradId() {
+    public Long getCradId() {
         return cradId;
     }
 
-    public void setCradId(Integer cradId) {
+    public void setCradId(Long cradId) {
         this.cradId = cradId;
     }
