+++ creditcard-interface/src/main/java/com/credit/pojo/TbContactInfo.java
@@ -9,33 +9,33 @@ import javax.persistence.Table;
 public class TbContactInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;//自增id
+    private Long id;//自增id
 
-    private Integer uid;//用户uid
+    private Long uid;//用户uid
 
     private String contactName;//联系人姓名
 
     private String contactPhone;//联系人电话
 
-    private Integer reWithApplicant;//与申请人关系（1为父子，2为母子，3为夫妻，4为亲戚，5为朋友，6为同事）
+    private Long reWithApplicant;//与申请人关系（1为父子，2为母子，3为夫妻，4为亲戚，5为朋友，6为同事）
 
     private String email;//电子邮箱
 
-    private Integer getCardMethod;//卡片领取方式（1为挂号信邮寄，2银行网点自取）
+    private Long getCardMethod;//卡片领取方式（1为挂号信邮寄，2银行网点自取）
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
 
@@ -55,11 +55,11 @@ public class TbContactInfo {
         this.contactPhone = contactPhone == null ? null : contactPhone.trim();
     }
 
-    public Integer getReWithApplicant() {
+    public Long getReWithApplicant() {
         return reWithApplicant;
     }
 
-    public void setReWithApplicant(Integer reWithApplicant) {
+    public void setReWithApplicant(Long reWithApplicant) {
         this.reWithApplicant = reWithApplicant;
     }
 
@@ -71,11 +71,11 @@ public class TbContactInfo {
         this.email = email == null ? null : email.trim();
     }
 
-    public Integer getGetCardMethod() {
+    public Long getGetCardMethod() {
         return getCardMethod;
     }
 
-    public void setGetCardMethod(Integer getCardMethod) {
+    public void setGetCardMethod(Long getCardMethod) {
         this.getCardMethod = getCardMethod;
     }
 }
